Cleaned up abstract test code and added groupby tests
