+++ src/test/resources/classificationDef/gkgToneQuantile.json
