+++ src/test/resources/classificationDef/objectIdGeometricInterval.json
