+++ src/test/resources/classificationDef/scoreEqualInterval.json
