+++ src/test/resources/classificationDef/scoreNaturalBreaks.json
