--- src/test/resources/convert/dateStats.json
@@ -1,14 +1,0 @@
-{
-  "params" : {
-    "id" : "Restaurants",
-    "layer" : 0,
-    "method" : "query"
-  },
-  "query" : {
-    "outStatistics": "[
-      {\"statisticType\":\"count\",\"onStatisticField\":\"OBJECTID\",\"outStatisticFieldName\":\"count\"},
-      {\"statisticType\":\"min\",\"onStatisticField\":\"Inspection\",\"outStatisticFieldName\":\"minimum_score\"},
-      {\"statisticType\":\"max\",\"onStatisticField\":\"Inspection\",\"outStatisticFieldName\":\"maximum_score\"}
-    ]"
-  }
-}
