--- src/test/resources/convert/groupByCityWithFilterByCity.json
@@ -1,13 +1,0 @@
-{
-	"params" : {
-    "id" : "Restaurants",
-    "layer" : 0,
-		"method" : "query"
-	},
-	"query" : {
-		"where" : "City='FONTANA' OR City='YUCCA VALLEY'",
-		"groupByFieldsForStatistics": "City",
-		"orderByFields": "objectid_count DESC",
-		"outStatistics": "[ {\"statisticType\":\"count\",\"onStatisticField\":\"OBJECTID\",\"outStatisticFieldName\":\"objectid_count\"} ]"
-	}
-}
