--- src/test/resources/convert/groupbyCity.json
@@ -1,13 +1,0 @@
-{
-	"params" : {
-    "id" : "Restaurants",
-    "layer" : 0,
-		"method" : "query"
-	},
-	"query" : {
-		"groupByFieldsForStatistics": "City",
-		"orderByFields": "objectid_count DESC",
-		"where": "1=1",
-		"outStatistics": "[ {\"statisticType\":\"count\",\"onStatisticField\":\"OBJECTID\",\"outStatisticFieldName\":\"objectid_count\"} ]"
-	}
-}
