--- src/test/resources/convert/groupbyCityScore.json
@@ -1,13 +1,0 @@
-{
-	"params" : {
-    "id" : "Restaurants",
-    "layer" : 0,
-		"method" : "query"
-	},
-	"query" : {
-		"groupByFieldsForStatistics": "City, Score",
-		"orderByFields": "City ASC, Score DESC",
-		"where": "1=1",
-		"outStatistics": "[ {\"statisticType\":\"count\",\"onStatisticField\":\"OBJECTID\",\"outStatisticFieldName\":\"objectid_count\"} ]"
-	}
-}
