--- src/test/resources/convert/statsOnly.json
@@ -1,16 +1,0 @@
-{
-	"params" : {
-    "id" : "Restaurants",
-    "layer" : 0,
-		"method" : "query"
-	},
-	"query" : {
-		"orderByFields": "average_score DESC",
-		"outStatistics": "[
-			{\"statisticType\":\"count\",\"onStatisticField\":\"OBJECTID\",\"outStatisticFieldName\":\"count\"},
-			{\"statisticType\":\"avg\",\"onStatisticField\":\"Score\",\"outStatisticFieldName\":\"average_score\"},
-			{\"statisticType\":\"min\",\"onStatisticField\":\"Score\",\"outStatisticFieldName\":\"minimum_score\"},
-			{\"statisticType\":\"max\",\"onStatisticField\":\"Score\",\"outStatisticFieldName\":\"maximum_score\"}
-		]"
-	}
-}
